requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
jinja2==3.1.3
python-dotenv==1.0.1

python-multipart==0.0.6
groq
